+++ app/views/admin/events.php
@@ -233,9 +233,7 @@
 >
     <i class="fas fa-edit"></i>
 </button>
-                                    <button class="btn-action-table delete" onclick="deleteEvent(<?php echo $event['id']; ?>)" title="Delete">
-                                        <i class="fas fa-trash"></i>
-                                    </button>
+                                    <!-- No delete button for upcoming events -->
                                 </td>
                             </tr>
                             <?php endforeach; ?>
@@ -299,9 +297,18 @@
                                     <button class="btn-action-table view" onclick="viewEvent(<?php echo $event['id']; ?>)" title="View">
                                         <i class="fas fa-eye"></i>
                                     </button>
-                                    <button class="btn-action-table delete" onclick="deleteEvent(<?php echo $event['id']; ?>)" title="Delete">
-                                        <i class="fas fa-trash"></i>
-                                    </button>
+                                    <?php if (!empty($event['can_delete'])): ?>
+                                        <button class="btn-action-table delete" onclick="deleteEvent(<?php echo $event['id']; ?>)" title="Delete">
+                                            <i class="fas fa-trash"></i>
+                                        </button>
+                                    <?php else: ?>
+                                        <button class="btn-action-table delete disabled" 
+                                                title="Can only delete events 6 months after end date" 
+                                                style="opacity: 0.5; cursor: not-allowed;" 
+                                                disabled>
+                                            <i class="fas fa-trash"></i>
+                                        </button>
+                                    <?php endif; ?>
                                 </td>
                             </tr>
                             <?php endforeach; ?>
